issues #59 issue #111 change location of stat calculation, change skills fragment to not take bundle arguments but instead have data set from the character detail activity. ability scores are now scores from stat generation (show added modifiers). reworked skill fragment and recycler to show values as + or -. skills fragment passes value of the skill to the recycler.
