+++ app/src/main/java/edu/ycp/cs482/iorc/Activities/CharacterDetailActivity.java
@@ -25,7 +25,12 @@ import com.google.gson.Gson;
 
 import org.jetbrains.annotations.NotNull;
 
+import java.io.Serializable;
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+import java.util.ArrayList;
 import java.util.HashMap;
+import java.util.List;
 
 import edu.ycp.cs482.iorc.CharacterVersionQuery;
 
@@ -35,9 +40,12 @@ import edu.ycp.cs482.iorc.Fragments.CharacterPanels.MagicFragment;
 import edu.ycp.cs482.iorc.Fragments.CharacterPanels.SkillsFragment;
 import edu.ycp.cs482.iorc.Fragments.MasterFlows.ItemDetailFragment;
 import edu.ycp.cs482.iorc.R;
+import edu.ycp.cs482.iorc.VersionSheetQuery;
 import edu.ycp.cs482.iorc.fragment.CharacterData;
+import edu.ycp.cs482.iorc.fragment.ClassData;
 import edu.ycp.cs482.iorc.fragment.ItemData;
-
+import edu.ycp.cs482.iorc.fragment.RaceData;
+import edu.ycp.cs482.iorc.fragment.VersionSheetData;
 
 
 /**
@@ -54,9 +62,11 @@ public class CharacterDetailActivity extends AppCompatActivity implements Equipm
     private static final String V_DATA = "VERSION_DATA";
     private CharacterVersionQuery.GetCharactersByVersion mItem;
     private String CHARCTER_ID = "";
+    private VersionSheetQuery.GetVersionSheet versionData;
+    private HashMap<String, Double> charStatMap = new HashMap<>();
+    private List<VersionSheetData.Stat> skillList = new ArrayList<>();
+    private HashMap<String, Double> skillValueMap = new HashMap<>();
     private static final String CREATION_DATA = "CREATION_DATA";
-    public static final String ITEM_ID = "item_id";
-    public static final String MAP_ID = "map_id";
     private HashMap<String, String> creationData;
 
 
@@ -83,22 +93,36 @@ public class CharacterDetailActivity extends AppCompatActivity implements Equipm
             actionBar.setDisplayHomeAsUpEnabled(true);
         }
 
-        //Display Character name on Toolbar
-        Bundle char_Arguments = new Bundle();
-        char_Arguments.putSerializable(V_DATA, getIntent().getSerializableExtra(V_DATA));
-        char_Arguments.putString(CharacterDetailFragment.ARG_ITEM_ID,
-                getIntent().getStringExtra(CharacterDetailFragment.ARG_ITEM_ID));
-        char_Arguments.putSerializable(CharacterDetailFragment.ARG_MAP_ID,
-                getIntent().getSerializableExtra(CharacterDetailFragment.ARG_MAP_ID));
-        CharacterDetailFragment char_Fragment = new CharacterDetailFragment();
-        char_Fragment.setArguments(char_Arguments);
+        Intent extra = getIntent();
 
-        HashMap<String, String> charMap =(HashMap<String, String>)char_Arguments.getSerializable(CharacterDetailFragment.ARG_MAP_ID);
-        String charObj = "";
-        if(charMap != null){
-            charObj = charMap.get(char_Arguments.getString((CharacterDetailFragment.ARG_ITEM_ID)));
+        //receive map from character list activity
+        if (extra.getSerializableExtra(CharacterDetailFragment.ARG_ITEM_ID) != null&& extra
+                .getSerializableExtra(CharacterDetailFragment.ARG_MAP_ID) != null) {
+            //Log.d("CHAR_ARGUMENTS", getArguments().toString());
+            HashMap<String, String> charMap =(HashMap<String, String>)extra.
+                    getSerializableExtra(CharacterDetailFragment.ARG_MAP_ID);
+            String charObj = "";
+            if(charMap != null){
+                //Log.d("CHAR_OBJ", charMap.get(bundle.getString(ARG_ITEM_ID)));
+                charObj = charMap.get(extra.getStringExtra(CharacterDetailFragment.ARG_ITEM_ID));
+            }
+            mItem = (new Gson()).fromJson(charObj, CharacterVersionQuery.GetCharactersByVersion.class);
+            //Log.d("mItem CHECK: ", "" + mItem);
         }
-        mItem = (new Gson()).fromJson(charObj, CharacterVersionQuery.GetCharactersByVersion.class);
+
+        if(extra.getSerializableExtra(V_DATA) != null){
+            HashMap<String, String> vDataMap = (HashMap<String, String>)getIntent().getSerializableExtra(V_DATA);
+            if(vDataMap != null){
+                versionData = (new Gson()).fromJson(vDataMap
+                        .get(V_DATA), VersionSheetQuery.GetVersionSheet.class);
+            }
+        }
+
+        if(versionData != null){
+            generateCharacterStats();
+        }
+
+        //Display Character name on Toolbar
         CollapsingToolbarLayout appBarLayout = findViewById(R.id.toolbar_layout);
         if (appBarLayout != null && mItem != null) {
             appBarLayout.setTitle(mItem.fragments().characterData().name());
@@ -122,6 +146,7 @@ public class CharacterDetailActivity extends AppCompatActivity implements Equipm
             arguments.putString(CharacterDetailFragment.ARG_ITEM_ID,
                     CHARCTER_ID);
             arguments.putSerializable(V_DATA, getIntent().getSerializableExtra(V_DATA));
+            arguments.putSerializable(CharacterDetailFragment.ARG_CHAR_STAT_DATA, charStatMap);
             arguments.putSerializable(CharacterDetailFragment.ARG_MAP_ID,
                     getIntent().getSerializableExtra(CharacterDetailFragment.ARG_MAP_ID));
             CharacterDetailFragment fragment = new CharacterDetailFragment();
@@ -158,22 +183,6 @@ public class CharacterDetailActivity extends AppCompatActivity implements Equipm
                 confirmDeleteBox();
                 break;
         }
-        //int id = item.getItemId();
-        //if (id == android.R.id.home) {
-        // This ID represents the Home or Up button. In the case of this
-        // activity, the Up button is shown. For
-        // more details, see the Navigation pattern on Android Design:
-        //
-        // http://developer.android.com/design/patterns/navigation.html#up-vs-back
-        //
-        //  navigateUpTo(new Intent(this, CharacterListActivity.class));
-        //return true;
-        //}
-        //else if(id == R.id.deleteCheck){
-        //  Intent deleteIntent = new Intent(CharacterDetailActivity.this, DeleteCheckActivity.class);
-        //startActivity(deleteIntent);
-        //}
-
         return super.onOptionsItemSelected(item);
     }
 
@@ -188,6 +197,7 @@ public class CharacterDetailActivity extends AppCompatActivity implements Equipm
                     Log.d("V_DATA", arguments.toString());
                     arguments.putString(CharacterDetailFragment.ARG_ITEM_ID,
                             getIntent().getStringExtra(CharacterDetailFragment.ARG_ITEM_ID));
+                    arguments.putSerializable(CharacterDetailFragment.ARG_CHAR_STAT_DATA, charStatMap);
                     arguments.putSerializable(CharacterDetailFragment.ARG_MAP_ID,
                             getIntent().getSerializableExtra(CharacterDetailFragment.ARG_MAP_ID));
                     CharacterDetailFragment fragment = new CharacterDetailFragment();
@@ -200,11 +210,9 @@ public class CharacterDetailActivity extends AppCompatActivity implements Equipm
                 case R.id.action_skills:
                     //TODO: Please use newInstance method if at all possible
                     Bundle skillArguments = new Bundle();
-                    skillArguments.putSerializable(V_DATA, getIntent().getSerializableExtra(V_DATA));
-                    //skillArguments.putString(SkillsFragment.ARG_ITEM_ID,
-                    //getIntent().getStringExtra(SkillsFragment.ARG_ITEM_ID));
-                    //skillArguments.putSerializable(SkillsFragment.ARG_MAP_ID, getIntent().getSerializableExtra(SkillsFragment.ARG_MAP_ID));
                     SkillsFragment skillFragment = new SkillsFragment();
+                    skillFragment.assignSkillList(skillList);
+                    skillFragment.assignSkillMap(skillValueMap);
                     skillFragment.setArguments(skillArguments);
                     getSupportFragmentManager().beginTransaction()
                             .replace(R.id.character_detail_container, skillFragment)
@@ -232,7 +240,7 @@ public class CharacterDetailActivity extends AppCompatActivity implements Equipm
                         invMap.put(String.valueOf(i), itemData);
                         Log.d("INVENTORY_ITEM", itemData);
                     }
-                    //TODO: inject character inventory into fragment
+                    //inject character inventory into fragment
                     EquipmentFragment fragment3 = EquipmentFragment.Companion.newInstance(invMap);
                     android.support.v4.app.FragmentTransaction fragmentTransaction3 = getSupportFragmentManager().beginTransaction();
                     fragmentTransaction3.replace(R.id.character_detail_container, fragment3, "FragmentName");
@@ -285,5 +293,209 @@ public class CharacterDetailActivity extends AppCompatActivity implements Equipm
         Intent intent = new Intent(this, ItemDetailActivity.class);
         intent.putExtra(ItemDetailFragment.ARG_ITEM, gsonItem);
         startActivity(intent);
+    }
+
+    public void generateCharacterStats(){
+        //TODO unspaghet
+
+        List<VersionSheetData.Stat> stats = versionData.fragments().versionSheetData().stats();
+        List<String> abilityList = new ArrayList<>();
+        List<String> modifierList = new ArrayList<>();
+        CharacterData charData = mItem.fragments().characterData();
+        HashMap<String, VersionSheetData.Stat> statObjMap = new HashMap<>();
+
+        Log.d("VERSION_DATA", versionData.fragments().versionSheetData().stats().toString());
+
+        CharacterData.Classql charClass = charData.classql();
+        CharacterData.Race charRace = charData.race();
+        CharacterData.AbilityPoints abilityScores = mItem.fragments().characterData().abilityPoints();
+
+        //loop through stats list and add to hashmap with the key being the name of the stat
+        for(int i = 0; i < stats.size(); i++){
+
+            //get version sheet stats and add each one to a map
+            VersionSheetData.Stat stat = stats.get(i);
+            String key = keyFilter(stat);
+            statObjMap.put(key, stat);
+            charStatMap.put(key, 0d);
+
+        }
+
+        Log.d("CHARACTER_ABILITIES", charStatMap.toString());
+
+        //add race values to map
+        for(int i = 0; i < charRace.fragments().raceData().modifiers().size(); i++){
+
+            //get racedata modifiers
+            RaceData.Modifier mod = charRace.fragments().raceData().modifiers().get(i);
+
+            //check if the modifier key is in the map if so add the value to the existing
+            //value in the map
+            if(charStatMap.containsKey(mod.key())){
+                Double mapItem = charStatMap.get(mod.key());
+                mapItem += mod.value();
+                charStatMap.put(mod.key(), mapItem);
+            }
+        }
+
+        //add class values to map
+        for(int i = 0; i < charClass.fragments().classData().modifiers().size(); i++){
+
+            //get racedata modifiers
+            ClassData.Modifier mod = charClass.fragments().classData().modifiers().get(i);
+
+            //check if the modifier key is in the map if so add the value to the existing
+            //value in the map
+            if(charStatMap.containsKey(mod.key())){
+                Double mapItem = charStatMap.get(mod.key());
+                mapItem += mod.value();
+                charStatMap.put(mod.key(), mapItem);
+            }
+        }
+
+        //add ability points from character data
+        for(Method m: abilityScores.getClass().getMethods()){
+            String abilKey = m.getName().replaceAll("_", "");
+            if(charStatMap.containsKey(abilKey)){
+                Log.d("ADD_FROM_METHOD", abilKey);
+                try {
+                    long abilityVal = (long) m.invoke(abilityScores);
+                    double val = charStatMap.get(abilKey);
+                    val += abilityVal;
+                    abilityList.add(abilKey);
+                    charStatMap.put(abilKey, val);
+                } catch (IllegalAccessException e) {
+                    e.printStackTrace();
+                } catch (InvocationTargetException e) {
+                    e.printStackTrace();
+                }
+            }
+        }
+
+        //create ability modifiers
+        for(String key : statObjMap.keySet()){
+            //Log.d("KEY", key);
+            VersionSheetData.Stat statObj = statObjMap.get(key);
+            List<VersionSheetData.Modifier> statMods = statObj.modifiers();
+
+            //modifier values
+            Double abilModVal = null;
+            Double mulVal = null;
+
+            if(statMods != null){
+                //get the ability modifier for each ability
+                for(VersionSheetData.Modifier mod : statMods){
+                    if(abilityList.contains(mod.key())){
+                        Log.d("KEY", mod.key());
+                        abilModVal = charStatMap.get(mod.key());
+                        abilModVal += mod.value();
+                        modifierList.add(key);
+                    }
+                    if(mod.key().equals("*")){
+                        mulVal = mod.value();
+                    }
+                }
+
+                //calculate ability modifier and add to the modifier value in the map
+                if(abilModVal != null && mulVal != null){
+                    abilModVal *= mulVal;
+                    charStatMap.put(key, Math.floor(abilModVal));
+                }
+
+            }
+
+        }
+
+        for(String key : statObjMap.keySet()){
+            VersionSheetData.Stat statObj = statObjMap.get(key);
+            List<VersionSheetData.Modifier> statMods = statObj.modifiers();
+
+            Double skillValue;
+
+            if(statMods != null){
+                String statName = statObj.name();
+                for(VersionSheetData.Modifier mod : statMods){
+
+                    //create skill values
+                    if(modifierList.contains(mod.key())){
+                        //Log.d("MOD_KEY", mod.key() + statName);
+                        skillValue = charStatMap.get(statName.toLowerCase());
+                        skillValue += charStatMap.get(mod.key());
+                        //Log.d("SKILL_NAME", statName.toLowerCase());
+                        charStatMap.put(statName.toLowerCase(), skillValue);
+                    }
+
+                    //create defense values
+                    if(mod.key().contains("||")){
+                        String[] options = mod.key().split("\\|\\|");
+
+                        //Log.d("OPTIONS", options[0] + " " + options[1]);
+                        //Log.d("DEFENSE", statName.toLowerCase());
+
+                        String highestMod = selectDefMod(options);
+                        //Log.d("DEF_MOD", highestMod);
+                        Double value =  charStatMap.get(highestMod) + 10;
+                        //Log.d("DEF_VAL", value.toString());
+                        charStatMap.put(statName.toLowerCase(), value);
+
+                    }
+                }
+                //add skills to skill map
+                if(statObj.skill()){
+                    skillValueMap.put(key, charStatMap.get(key));
+                    skillList.add(statObj);
+                }
+            }
+
+
+        }
+        //Log.d("CHARACTER_ABILITIES", charStatMap.toString());
+        //log map values
+
+        printMapValues(charStatMap);
+        Log.d("PRINT_SKILLS", "HERE");
+        printMapValues(skillValueMap);
+
+    }
+
+    public String keyFilter(VersionSheetData.Stat stat){
+        String statName = stat.name();
+        String statDescription = stat.description();
+        String keyName;
+
+        //abilities have themselves as the description
+        //if a stat is an ability get the first 3 characters of the string
+        //then make the substring the string name
+        if(statName.equals(statDescription)){
+            statName = statName.substring(0, 3);
+        }
+
+        //make sure the key is all lowercase
+        keyName = statName.toLowerCase();
+
+        return keyName;
+    }
+
+    public String selectDefMod(String[] options){
+
+        String highest = "";
+        Double highestVal = null;
+
+        for(String option : options){
+            Double value = charStatMap.get(option);
+            if(highestVal == null || highestVal < value){
+                highestVal = value;
+                highest = option;
+            }
+        }
+
+        return highest;
+    }
+
+    //print a map with the key and corresponding value to make it easier to read (for debugging/dev)
+    public void printMapValues(HashMap<String, Double> map){
+        for(String key : map.keySet()){
+            Log.d("MAP_VALUE", key + " " + map.get(key).toString());
+        }
     }
 }
